main/EulerEye-master3/requirements.txt
botocore>=1.22.5
boto3
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.11.0
docutils==0.14
graphviz==0.8.4
idna==3.3
imageio==2.10.1
jmespath==0.10.0
kiwisolver
lockfile==0.11.0
matplotlib==3.4.3
mxnet==1.8.0.post0
networkx==2.6.3
numpy==1.21.3
opencv-python==4.5.4.58
opencv-python-headless==4.5.3.56
pandas==1.3.4
Pillow==8.4.0
pyparsing==3.0.4
pystache==0.5.4
python-daemon==2.2.3
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
requests==2.26.0
s3transfer==0.5.0
scikit-image==0.18.3
scipy==1.7.1
Shapely==1.8.0
simplejson==3.2.0
six==1.16.0
tifffile==2021.10.12
torch==1.10.0
torchvision==0.11.1
typing-extensions==3.10.0.2
urllib3==1.26.7
